Ping non-DHCP containers before bootstrapping
Since we will have an IP to look for if we used a DNS provider,
instead of just sleeping and hoping the cotainer comes up, we will
ping it.

If we don't get a response and time out then we assume its bad and
bail on bootstrapping.

Also only use networking templates for non-DHCP containers as the
default configs in the images should just work for DHCP (to be
revisited later)