Move ALL ws messages to buffer, lock map access, add management func
I forgot to move the various ws writes that happened after we called a
handler to the buffer which could result in concurrent writes to a
socket and crash.  Now we only write from the consume function making
them much easier to control.  I pity the fool that uses this and we
can't create a buffer for (will have to look into that later).

Our message buffer map can be read / written from multiple processes so
locking access to that with a mutex

Added our management function to clean out old buffers